Deprecate QXmppStarttlsPacket
